chore: Move SonarQube analysis to CI/CD workflow
The deploy and analyze workflows are currently separate, delete the
analyze workflow and move the quality checks to the deploy workflow
instead.
Allow the deploy workflow to run on all pushes, but stop dockerization
if the branch is not master.
Rename deploy.yml to ci-cd-workflow.yml to better describe its dual
responsibility.

TISNEW-3848