fix: Updating a trainee skeleton causes a NPE
The update mapper for PersonalDetails attempts to update the existing
PersonalDetails for the TraineeProfile. In the case of a skeleton
profile being created this will be null and the update will throw a
NullPointerException.
Update the mappers in such a way that it will automatically create and
set the PersonalDetails field in the TraineeProfile as part of the
update if it is null.

TISNEW-3833