fix: only assemble fatjar
The bump to org.springframework.boot 2.5+ means that running gradlew
assemble creates both the fatjar and the jar without dependencies
(*-plain.jar). The problem is that the Dockerfile attempts to copy both
to app.jar, so Dockerize fails with:
Step 3/4 : COPY build/libs/*.jar app.jar
When using COPY with more than one source file, the destination must be a directory and end with a /
Error: Process completed with exit code 1.