feat: add unique index on traineeTisId
This might be straying a bit into Mongock territory, so very happy to
refactor to use that if desired (it was actually how I implemented it at
first, but this way seems more concise).

Note that if Mongock is used, the MongoConfiguration should have
the ensureIndex() removed for traineeTisId, since this can throw an
exception if the existing index on that field has a different unique
value or has a different name, before Mongock can rework things.

TIS21-2923: Duplicate profile for trainee