fix: return 201 Created from credential endpoint
The credential should return a 201 with the redirect URI in the Location
header so that the client app can perform the redirect manually, browser
redirect is not possible due to CORS restrictions on redirecting between
different domains.

Refactor the DspConfigurationProperties to use Lombok `@Data` to
generate the boilerplate getters/setters.

TIS21-3685