feat: make placement and prog mem signed DTOs
Placement and ProgrammeMembership DTOs must be signed so that their data
can be verified.
Create a SignedDto interface that both PlacementDto and
ProgrammeMembershipDto will extend. The interface will declare a setter
for a new Signature object.
Create a Signature object to include on signed DTOs, the object will
have `signedAt` `validUntil` and `hmac` fields to record the signature
hash and timestamps.

TIS21-3681
TIS21-3691