fix: skip processing after finding 1st legacy COJ
If the legacy Programme Membership data is structured incorrectly, i.e.
as two PMs instead of one with multiple curricula, it is possible for a
valid COJ to not be copied over to the new UUID PM.
This occurs when the first returned PM has a COJ but subsequent matching
PMs do not, the later `null` values overwrite the originally assigned
COJ.

Add a null check and a break to ensure that we only set COJ values that
are non-null and exit the loop early if we've found a COJ already.

TIS21-2399