refactor: the personal details update event
The data structure for the PersonalDetailsEvent does not currently match
the data that is sent from tis-trainee-sync.

Update the event DTO to more closely reflect the structure of the
incoming message data.
Add SQS config to ensure that messages are read correctly.

Add an alias to the PersonalDetailsDto to ensure that `owner` can be
correctly mapped to `personOwner`.

TIS21-4086
TIS21-4076