feat: add Other Sites to placement records
Create a Site entity and SiteDto to standardize the fields used to
represent sites.
Update the existing Placement entity and PlacementDto to include a
collectioin of Other Sites as well as use the Site entity/dto for the
existing site fields. Those existing fields will be unwrapped
automatically so that the existing DTO/DB structures are unchanged.

TIS21-4970
TIS21-4396