feat: create NTN generator
Create an NtnGenerator service which can populate the NTN a profile's
programme memberships.

The logic follows the existing Tableau/SQL implementation with the
exception of the limitation on more than four specialties being removed.
As a result the NTN will include as many specialties as exist and
sub-specialties will add use the `.XYZ` notation instead of only the
final one.

TIS21-6175
TIS21-6182