feat: generate NTNs when retrieving trainee data
The NTN should eventually be generated and stored against the PM on
create/update and on a daily schedule to account for the starting and
ending of curricula.
However, temporarily it can be generated when retrieving the trainee
profile so that it is available to trainees while the rest of the work
is being done to make it available via metabase.

TIS21-6183
TIS21-6175