fix: NTN generation with partial profile data
Ensure that the NTN generation fails gracefully and does not throw and
exception getting the trainee profile data if the training pathway is
null. This scenario exists for a tiny number of TIS records but until
data is resynced the training pathway will be missing for all profiles
and should not throw an exception.

Update the curricula filter to ensure that curricula with no specialty
codes are excluded. This avoids an issue with NTNs being generated like
`LDN/null-null/1234567/D`

TIS21-6175
TIS21-6182