feat: return validation errors in REST response
Create a custom RestResponseEntityExceptionHandler to handle the
inclusion of validation failure messages in API responses. Follow the
ProblemDetails structure with JSON Pointers used for request body based
validation.

Example:
```json
{
  "type": "about:blank",
  "title": "Validation failure",
  "status": 400,
  "instance": "/trainee/api/basic-details/gmc-number",
  "errors": [
    {
      "pointer": "#/gmcNumber",
      "detail": "must be 7 digits"
    }
  ]
}
```

Refactor the update GMC endpoint to accept a request body instead of a
path variable containing the GMC number. Add validation to ensure that
the GMC number is valid and the GMC status is not provided by the
caller.

TIS21-6292
TIS21-6347