fix: show all programmes
Currently PAST programmes are filtered out, but they now need to be
shown.
At the same time an attempt to filter placements in the same way fails
because the placement status is only ever `CURRENT` or `null`.

Update TraineeProfileResource to remove the filtering of past data.
Update TraineeProfileService to remove the functions that perform the
filtering, along with a filter for deprecated programmes which is no
longer used.

TIS21-6386
TIS21-6387
TIS21-6388