feat: forbid CCT requests with no trainee ID
Previously each endpoint that required a trainee ID checked for the ID
and returned `400 Bad Request` if the ID was missing. Now there is a
centralised place for managing the `custom:tisId` token property the
trainee ID check can be done in a single place.
However, several existing endpoints are private/internal and do not
receive a token at all.

Update the TraineeIdentityInterceptor to reject only requests to the CCT
controller if the trainee ID is missing. In such cases a `403 Forbidden`
should be returned.
The interceptor can take over verification for other endpoints once the
unauthorised endpoints are refactored to allow sensible filtering based
on the request URI.

TIS21-6544
TIS21-6623