feat(ltft): always enable for beta participants
Members of the `beta-participant` group should always have the LTFT
feature enabled, regardless of having a qualifying programme membership.
Update the trainee identity to include the Cognito groups, use the
groups to determine beta group membership when checking valid feature
flags.

Refactor the FeaturesProperties to allow multiple tranches to be
configured with individual start dates and deanery lists.
The feature service should combine all started tranches when determining
whether there is a valid PM or not.

Update the TraineeIdentityInterceptor to relax the trainee ID
requirement on the feature flag endpoint.
Features will still be disabled via the FeatureService if there is no
trainee ID, but the hard requirement in the interceptor causes
authentication failures with ambiguous error messages.

TIS21-6703
TIS21-6908